% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensureProductionInputs.R
\name{ensureProductionInputs}
\alias{ensureProductionInputs}
\title{Ensure Production Inputs}
\usage{
ensureProductionInputs(data, processingParameters)
}
\arguments{
\item{data}{A data.table containing the data.}

\item{processingParameters}{A list containing the parameters to be used in
the processing algorithms.  See ?defaultProcessingParameters for a starting
point.}
}
\description{
This function is designed to ensure that the provided dataset is valid.  In
particular, it coerces column types: all values are coerced to numeric (
instead of integer, which can cause problems) and all flags are coerced to
character (instead of logical, which occurs if the flag is set to NA).
Also, it ensures data is a data.table.
}

