% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getProductionDomainData.R
\name{getProductionDomainData}
\alias{getProductionDomainData}
\title{Get Production Domain Data}
\usage{
getProductionDomainData(key)
}
\arguments{
\item{key}{A DatasetKey object, typically as created by GetTestEnvironment.
See the argument with the same name in faosws::GetData.}
}
\value{
A data.table object containing the dataset of interest.
}
\description{
This function is designed to pull production data from the working system.
It is essentially a wrapper to the GetData function in faosws, but it
massages the data from that function slightly.
}

