% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/processProductionDomain.R
\name{processProductionDomain}
\alias{processProductionDomain}
\title{This is a wrapper for all the data manipulation step before the
preparation of the imputation.}
\usage{
processProductionDomain(data, processingParameters)
}
\arguments{
\item{data}{The data}

\item{processingParameters}{A list of the parameters for the production
processing algorithms.  See defaultProductionParameters() for a starting
point.}
}
\description{
This is a wrapper for all the data manipulation step before the
preparation of the imputation.
}

