% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/saveProductionData.R
\name{saveProductionData}
\alias{saveProductionData}
\title{Save Production Data}
\usage{
saveProductionData(data)
}
\arguments{
\item{data}{The data.table object containing the seed data to be written to
the database.}
}
\value{
No R objects are returned, as this functions purpose is solely to
write to the database.
}
\description{
This function takes the a seed data dataset and saves it back to the
database.
}

